+++ rabix-bindings/src/main/java/org/rabix/bindings/ProtocolFilePathMapper.java
@@ -1,14 +1,10 @@
 package org.rabix.bindings;
 
-import java.io.File;
-
 import org.rabix.bindings.filemapper.FileMapper;
 import org.rabix.bindings.model.Job;
 
-public interface ProtocolPreprocessor {
+public interface ProtocolFilePathMapper {
 
-  Job preprocess(final Job job, final File workingDir) throws BindingException;
-  
   Job mapInputFilePaths(final Job job, final FileMapper fileMapper) throws BindingException;
   
   Job mapOutputFilePaths(final Job job, final FileMapper fileMapper) throws BindingException;
