+++ rabix-bindings/src/main/java/org/rabix/bindings/protocol/draft2/Draft2FilePathMapper.java
@@ -1,45 +1,19 @@
 package org.rabix.bindings.protocol.draft2;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.Map;
 
-import org.apache.commons.io.FileUtils;
 import org.rabix.bindings.BindingException;
-import org.rabix.bindings.ProtocolPreprocessor;
+import org.rabix.bindings.ProtocolFilePathMapper;
 import org.rabix.bindings.filemapper.FileMapper;
 import org.rabix.bindings.model.Job;
 import org.rabix.bindings.protocol.draft2.bean.Draft2Job;
 import org.rabix.bindings.protocol.draft2.helper.Draft2JobHelper;
 import org.rabix.bindings.protocol.draft2.processor.Draft2PortProcessor;
 import org.rabix.bindings.protocol.draft2.processor.Draft2PortProcessorException;
-import org.rabix.bindings.protocol.draft2.processor.callback.Draft2PortProcessorHelper;
 import org.rabix.bindings.protocol.draft2.processor.callback.FilePathMapProcessorCallback;
-import org.rabix.common.json.BeanSerializer;
 
-public class Draft2Preprocessor implements ProtocolPreprocessor {
+public class Draft2FilePathMapper implements ProtocolFilePathMapper {
 
-  private static final String JOB_FILE = "job.json";
-  
-  @Override
-  public Job preprocess(final Job job, final File workingDir) throws BindingException {
-    Draft2Job draft2Job = Draft2JobHelper.getDraft2Job(job);
-    Draft2PortProcessorHelper portProcessorHelper = new Draft2PortProcessorHelper(draft2Job);
-    try {
-      File jobFile = new File(workingDir, JOB_FILE);
-      String serializedJob = BeanSerializer.serializePartial(Draft2JobHelper.getDraft2Job(job));
-      FileUtils.writeStringToFile(jobFile, serializedJob);
-      
-      Map<String, Object> inputs = job.getInputs();
-      inputs = portProcessorHelper.setFileSize(inputs);
-      inputs = portProcessorHelper.loadInputContents(inputs);
-      inputs = portProcessorHelper.stageInputFiles(inputs, workingDir);
-      return Job.cloneWithInputs(job, inputs);
-    } catch (Draft2PortProcessorException | IOException e) {
-      throw new BindingException(e);
-    }
-  }
-  
   @Override
   public Job mapInputFilePaths(final Job job, final FileMapper fileMapper) throws BindingException {
     Draft2Job draft2Job = Draft2JobHelper.getDraft2Job(job);
@@ -65,5 +39,5 @@ public class Draft2Preprocessor implements ProtocolPreprocessor {
       throw new BindingException(e);
     }
   }
-  
+
 }
